==> scripts/1.3-HotspotMarchOctober.Rmd <==
---
title: "Subset Annual Modis for March-October"
author: "Anna Talucci"
date: "2/10/2021"
output: html_document
---

# Overview

This script subsets MODIS hotspot data for March thru October each year from 2001-2020.

Data acquired from [FIRMS](https://firms.modaps.eosdis.nasa.gov/download/create.php)

Description of [Modis data](https://earthdata.nasa.gov/earth-observation-data/near-real-time/firms)

MODIS data is available 2001-present and was use instead of VIIRS, which launched October 2011 and has fire season data available 2012-present. Adding the data required more computing processing then available.

The data is subset March to October to capture wildfire activity. The point data is eventually convert to polygons in script 1.4.


# Clear workspace

```{r}
rm(list=ls()) 
```

# Packages

```{r include=FALSE}
library(sp)
library(sf)
library(rgdal) # geospatial data
library(raster)
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(spatialEco)
library(lubridate)
```

# Read in MODIS  data (2001-2020)

```{r}
hs = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2001.shp", "fire_archive_M6_2001") 
```

```{r M6-files-2001-11 }
hs2001 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2001.shp", "fire_archive_M6_2001") 
hs2002 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2002.shp", "fire_archive_M6_2002") 
hs2003 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2003.shp", "fire_archive_M6_2003") 
hs2004 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2004.shp", "fire_archive_M6_2004") 
hs2005 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2005.shp", "fire_archive_M6_2005") 

hs2006 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2006.shp", "fire_archive_M6_2006") 
hs2007 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2007.shp", "fire_archive_M6_2007") 
hs2008 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2008.shp", "fire_archive_M6_2008") 
hs2009 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2009.shp", "fire_archive_M6_2009") 
hs2010 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2010.shp", "fire_archive_M6_2010") 
hs2011 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2011.shp", "fire_archive_M6_2011") 
```

```{r M6-files-2012-2020 }
hs2012 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2012.shp", "fire_archive_M6_2012") 
hs2013 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2013.shp", "fire_archive_M6_2013") 
hs2014 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2014.shp", "fire_archive_M6_2014") 
hs2015 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2015.shp", "fire_archive_M6_2015") 

hs2016 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2016.shp", "fire_archive_M6_2016") 
hs2017 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2017.shp", "fire_archive_M6_2017") 
hs2018 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2018.shp", "fire_archive_M6_2018") 
hs2019 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2019.shp", "fire_archive_M6_2019") 
hs2020 = st_read("../data/FIRMS_Annual_PointData/fire_archive_M6_2020.shp", "fire_archive_M6_2020") 
```



# Subset from March to October for each year
```{r}
hs2001_month = hs2001 %>% filter(ACQ_DATE >= "2001/03/01" & ACQ_DATE <="2001/10/31")
hs2002_month = hs2002 %>% filter(ACQ_DATE >= "2002/03/01" & ACQ_DATE <="2002/10/31")
hs2003_month = hs2003 %>% filter(ACQ_DATE >= "2003/03/01" & ACQ_DATE <="2003/10/31")
hs2004_month = hs2004 %>% filter(ACQ_DATE >= "2004/03/01" & ACQ_DATE <="2004/10/31")
hs2005_month = hs2005 %>% filter(ACQ_DATE >= "2005/03/01" & ACQ_DATE <="2005/10/31")

hs2006_month = hs2006 %>% filter(ACQ_DATE >= "2006/03/01" & ACQ_DATE <="2006/10/31")
hs2007_month = hs2007 %>% filter(ACQ_DATE >= "2007/03/01" & ACQ_DATE <="2007/10/31")
hs2008_month = hs2008 %>% filter(ACQ_DATE >= "2008/03/01" & ACQ_DATE <="2008/10/31")
hs2009_month = hs2009 %>% filter(ACQ_DATE >= "2009/03/01" & ACQ_DATE <="2009/10/31")
hs2010_month = hs2010 %>% filter(ACQ_DATE >= "2010/03/01" & ACQ_DATE <="2010/10/31")

hs2011_month = hs2011 %>% filter(ACQ_DATE >= "2011/03/01" & ACQ_DATE <="2011/10/31")
```

```{r}

hs2012_month = hs2012 %>% filter(ACQ_DATE >= "2012/03/01" & ACQ_DATE <="2012/10/31")
hs2013_month = hs2013 %>% filter(ACQ_DATE >= "2013/03/01" & ACQ_DATE <="2013/10/31")
hs2014_month = hs2014 %>% filter(ACQ_DATE >= "2014/03/01" & ACQ_DATE <="2014/10/31")
hs2015_month = hs2015 %>% filter(ACQ_DATE >= "2015/03/01" & ACQ_DATE <="2015/10/31")

hs2016_month = hs2016 %>% filter(ACQ_DATE >= "2016/03/01" & ACQ_DATE <="2016/10/31")
hs2017_month = hs2017 %>% filter(ACQ_DATE >= "2017/03/01" & ACQ_DATE <="2017/10/31")
hs2018_month = hs2018 %>% filter(ACQ_DATE >= "2018/03/01" & ACQ_DATE <="2018/10/31")
hs2019_month = hs2019 %>% filter(ACQ_DATE >= "2019/03/01" & ACQ_DATE <="2019/10/31")
hs2020_month = hs2020 %>% filter(ACQ_DATE >= "2020/03/01" & ACQ_DATE <="2020/10/31")
```

# Write to shapefile

```{r eval=FALSE, include=FALSE}
st_write(hs2001_month, "../outputs/M6-march-october/M6_mar-oct_2001.shp", driver="ESRI Shapefile")
st_write(hs2002_month, "../outputs/M6-march-october/M6_mar-oct_2002.shp", driver="ESRI Shapefile")
st_write(hs2003_month, "../outputs/M6-march-october/M6_mar-oct_2003.shp", driver="ESRI Shapefile")
st_write(hs2004_month, "../outputs/M6-march-october/M6_mar-oct_2004.shp", driver="ESRI Shapefile")
st_write(hs2005_month, "../outputs/M6-march-october/M6_mar-oct_2005.shp", driver="ESRI Shapefile")
st_write(hs2006_month, "../outputs/M6-march-october/M6_mar-oct_2006.shp", driver="ESRI Shapefile")
st_write(hs2007_month, "../outputs/M6-march-october/M6_mar-oct_2007.shp", driver="ESRI Shapefile")
st_write(hs2008_month, "../outputs/M6-march-october/M6_mar-oct_2008.shp", driver="ESRI Shapefile")
st_write(hs2009_month, "../outputs/M6-march-october/M6_mar-oct_2009.shp", driver="ESRI Shapefile")
st_write(hs2010_month, "../outputs/M6-march-october/M6_mar-oct_2010.shp", driver="ESRI Shapefile")
st_write(hs2011_month, "../outputs/M6-march-october/M6_mar-oct_2011.shp", driver="ESRI Shapefile")
```

```{r eval=FALSE, include=FALSE}
st_write(hs2012_month, "../outputs/M6-march-october/M6_mar-oct_2012.shp", driver="ESRI Shapefile")
st_write(hs2013_month, "../outputs/M6-march-october/M6_mar-oct_2013.shp", driver="ESRI Shapefile")
st_write(hs2014_month, "../outputs/M6-march-october/M6_mar-oct_2014.shp", driver="ESRI Shapefile")
st_write(hs2015_month, "../outputs/M6-march-october/M6_mar-oct_2015.shp", driver="ESRI Shapefile")
st_write(hs2016_month, "../outputs/M6-march-october/M6_mar-oct_2016.shp", driver="ESRI Shapefile")
st_write(hs2017_month, "../outputs/M6-march-october/M6_mar-oct_2017.shp", driver="ESRI Shapefile")
st_write(hs2018_month, "../outputs/M6-march-october/M6_mar-oct_2018.shp", driver="ESRI Shapefile")
st_write(hs2019_month, "../outputs/M6-march-october/M6_mar-oct_2019.shp", driver="ESRI Shapefile")
st_write(hs2020_month, "../outputs/M6-march-october/M6_mar-oct_2020.shp", driver="ESRI Shapefile")

```
