==> scripts/1.4-HotspotToPolgygon.Rmd <==
---
title: "Hotspot point data to polygon"
author: "Anna Talucci"
date: "2/6/2020"
output: html_document
---

# Overview
Convert Hotspot point data from MODIS to polygons. The purpose of this is to limit the extent of the vector function in GEE to delineate fire perimeters. This script utilizes data process in script 1.3. The script runs a single year at a time. Use find replace to change year and then run to process next year. Special code for 2003, 2019,and 2020, due to large data set and computational limits. 

Data acquired from [FIRMS](https://firms.modaps.eosdis.nasa.gov/download/create.php)

Description of [Modis data](https://earthdata.nasa.gov/earth-observation-data/near-real-time/firms).

MODIS available 2001-present.

Take all annual hotspot data and make polygons 2001-2020. We did not end up combining MODIS/VIIRS data from 2012-2020, because the data set became to large to process.
 
Special Notes:

1. Processing hotspot point data to polygon data
2. Everything process to Russia Equal Area Projection
3. 2003, 2019 & 2020 are processed separately across 3 ROIs because of computational limits. 

# Clear workspace

```{r}
rm(list=ls()) 
```

# Packages

```{r include=FALSE}
library(sp)
library(sf)
library(rgdal) # geospatial data
library(raster)
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(spatialEco)
library(lubridate)
library(nngeo) # remove holes in polygons; works with sf objects 
```

# Projections

[Equal Area projection for Russia](https://gis.stackexchange.com/questions/264093/projection-for-russia-map) is needed for buffering hotspot data.

The projection:
+proj=aea +lat_1=50 +lat_2=70 +lat_0=56 +lon_0=100 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs 

```{r}
ee_russia = "+proj=aea +lat_1=50 +lat_2=70 +lat_0=56 +lon_0=100 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs "
```

# Data
##  Hotspot Data (MODIS)
**Change year on data with find and replace**

```{r}
hs = readOGR("../data/M6-march-october/M6_mar-oct_2003.shp", "M6_mar-oct_2003") 
```
## Single Ecozone Polygon
All nine ecozones have been combined into a single polygon

```{r}
eco_ee = readOGR("../data/ecozones/eco_SinglePoly_ee.shp", "eco_SinglePoly_ee") 
```

## Three ROI to split large data
GEE ROI (Regions of interest) to split the processing into three spatial extents including the NE Poly 1 that covers all of northeaster Siberia, Eastern portion of E Poly 2 and the Western portion of E Poly 2 that cover everything west of the northeastern region.

```{r}
eco_nepoly1_ee = readOGR("../data/ecozones/eco_NEPoly1_ee.shp", "eco_NEPoly1_ee") 
```

```{r}
eco_E_EPoly2_1_ee = readOGR("../data/ecozones/eco_E_EPoly2_1_ee.shp", "eco_E_EPoly2_1_ee") 
```

```{r}
eco_W_EPoly2_2_ee =  readOGR("../data/ecozones/eco_W_EPoly2_2_ee.shp", "eco_W_EPoly2_2_ee") 
```


# Reproject to equal area Russia
```{r}
hs_ee = spTransform(hs, CRS(ee_russia))
```



# Process point to polygon All years except 2003, 2019, 2020
This code uses a single polygon that covers all eight ecozones

1. Select points within ecozone region
2. Buffer points by 3000 m overlapping polygons join
3. Remove any holes in the polygons
5. Convert to sf object
6. Write to shapefile

## point to Polygon
```{r}
eco_hs = hs_ee[eco_ee, ]
eco_hs_3k = gBuffer(eco_hs, width = 3000, capStyle="ROUND", joinStyle="ROUND") # Buffer by 3k
eco_hs_3k_nh = remove.holes(eco_hs_3k) # Remove holes in polygons

```

```{r}
hs_poly_sf = st_as_sf(eco_hs_3k_nh)
```

## Write to shapefile
```{r eval=FALSE, include=FALSE}
st_write(hs_poly_sf, "../outputs/hotspot-poly-EqualArea/hspoly_EA_2001.shp", driver="ESRI Shapefile")
```


# Process Point to Polygon for Large data year 2003, 2019 & 2020)

Split the region of interest (ROI; single polygon) into 3 ROIs.  Process the point to polygon on the 3 ROIs and then combine into single shapefile.

1. Process point to polygon for each of three ecozone boundaries
2. Combine the polygons from the three ecozone boundaries into a single shapefile.

## For NE Poly 1
```{r}
eco1_hs = hs_ee[eco_nepoly1_ee, ]
eco1_hs_3k = gBuffer(eco1_hs, width = 3000, capStyle="ROUND", joinStyle="ROUND") # Buffer by 3k
eco1_hs_3k_nh = remove.holes(eco1_hs_3k) # Remove holes in polygons

eco1_hs_EA_sf = st_as_sf(eco1_hs_3k_nh)
```

## For eastern portion of E Poly 2
```{r}
eco2_hs = hs_ee[eco_E_EPoly2_1_ee, ]
eco2_hs_3k = gBuffer(eco2_hs, width = 3000, capStyle="ROUND", joinStyle="ROUND") # Buffer by 3k
eco2_hs_3k_nh = remove.holes(eco2_hs_3k) # Remove holes in polygons

eco2_hs_EA_sf = st_as_sf(eco2_hs_3k_nh)
```

## For western portion of E Poly 2
```{r}
eco3_hs = hs_ee[eco_W_EPoly2_2_ee, ]
eco3_hs_3k = gBuffer(eco3_hs, width = 3000, capStyle="ROUND", joinStyle="ROUND") # Buffer by 3k
eco3_hs_3k_nh = remove.holes(eco3_hs_3k) # Remove holes in polygons

eco3_hs_EA_sf = st_as_sf(eco3_hs_3k_nh)
```

## Combine Hotspot poly into one file for 2003, 2019 & 2020
1. Convert to sf object
2. `rbind` to combine into single file
3. Union to merge overlapping polygons
4. Cast as individual polygons
5. remove any holes that occur from processing
```{r}
eco1_sf = st_as_sf(eco1_hs_3k_nh)
eco2_sf = st_as_sf(eco2_hs_3k_nh)
eco3_sf = st_as_sf(eco3_hs_3k_nh) 

merge_polys = rbind(eco1_sf, eco2_sf, eco3_sf)

merge_polys_un = sf::st_union(merge_polys)

merge_polys_un_poly = sf::st_cast(merge_polys_un, "POLYGON")

merge_polys_un_poly_nh = nngeo::st_remove_holes(merge_polys_un_poly)
```

```{r}
plot(merge_polys_un_poly_nh)
```

## Write to shapefile
```{r eval=FALSE, include=FALSE}
st_write(merge_polys_un_poly_nh, "../outputs/hotspot-poly-EqualArea/hspoly_EA_2003.shp", driver="ESRI Shapefile")
```
