==> scripts/3.2-ShapefileAdvanceClean.Rmd <==
---
title: "Advanced Cleaning"
author: "Anna Talucci"
date: "5/13/2021"
output: html_document
---

# Clear workspace

```{r}
rm(list=ls()) 
```

# Overview

This script applies a negative buffer to counter act the buffer applied in script 3.1 and additional advanced processing steps to clean the polygons to that they are usable in Google Earth Engine (GEE).

Specific steps include:
1. Remove Holes
  + This removes any holes created in polygons when merging polygons in the previous script.
  + It is important to remove as holes because they can cause issues during geoprocessing.
2. Convert from sp to sf
3. Apply a smoothing function to reduce the pixelized appearance of the polygons.
4. Apply a simplify function to reduce the number of vertices in the polygon, which is an issue when using the polygons in GEE.



# Packages

```{r include=FALSE}
library(sp)
library(sf)
library(rgdal) # geospatial data
library(raster)
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(dplyr)
library(tidyr)
library(spatialEco)
library(ggplot2)
library(nngeo)
library(rmapshaper)
library(units)
library(smoothr)
```

# Read in shapefile
```{r}
vec = readOGR("../data/3.1-Clean/combine1_equalarea_2020.shp", "combine1_equalarea_2020") 
```

# Advance processing

## Apply negative buffer
```{r}
vec_nbuf = gBuffer(vec, width = -1000, capStyle="ROUND", joinStyle="ROUND")
```

## Convert from sp to sf
```{r}
vec_sf = st_as_sf(vec_nbuf)
```

## Apply smooth function

See [Smooth and Tidy Spatial Features](https://cran.r-project.org/web/packages/smoothr/vignettes/smoothr.html)

```{r}
vec_smooth <- smooth(vec_sf, method = "ksmooth", smoothness = 2)
```

## Apply Simplify function

Simplify with [mapshaper](https://cran.r-project.org/web/packages/rmapshaper/vignettes/rmapshaper.html).


```{r}
vec_simp = ms_simplify(vec_smooth, keep_shapes = FALSE, explode = TRUE)
```

# Write Shapefile
```{r}
st_write(vec_simp, "../outputs/3.2-AdvanceClean/nbuf_equalarea_2020.shp", driver="ESRI Shapefile") 
```
