==> scripts/3.5-CombineAllFireYears.Rmd <==
---
title: "Combine All Fire Years"
author: "Anna Talucci"
date: "5/21/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Clear workspace

```{r}
rm(list=ls()) 
```

# Overview

Combine individual fires year shapefiles into one shapefile for 2001-2020 plus some additional modification that include: 

1. Add fire size class. 
2. Remove fires less than 200 ha.
3. Remove fires with an average confidence of zero 
4. Calculate the number of days each fire burned

# Packages

```{r include=FALSE}
library(sp)
library(sf)
library(rgdal) # geospatial data
library(raster)
library(rgeos) # GIS Geometry
library(foreign) # deals with dbf files
library(tidyverse)
library(lubridate)
library(ids)
```


# Data

```{r}
f01 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2001.shp", "SiberiaFires_v2_2001") 
f02 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2002.shp", "SiberiaFires_v2_2002") 
f03 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2003.shp", "SiberiaFires_v2_2003") 
f04 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2004.shp", "SiberiaFires_v2_2004") 
f05 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2005.shp", "SiberiaFires_v2_2005") 
```

```{r}
f06 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2006.shp", "SiberiaFires_v2_2006") 
f07 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2007.shp", "SiberiaFires_v2_2007") 
f08 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2008.shp", "SiberiaFires_v2_2008") 
f09 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2009.shp", "SiberiaFires_v2_2009") 
f10 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2010.shp", "SiberiaFires_v2_2010") 
```

```{r}
f11 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2011.shp", "SiberiaFires_v2_2011") 
f12 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2012.shp", "SiberiaFires_v2_2012") 
f13 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2013.shp", "SiberiaFires_v2_2013") 
f14 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2014.shp", "SiberiaFires_v2_2014") 
f15 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2015.shp", "SiberiaFires_v2_2015") 
```

```{r}
f16 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2016.shp", "SiberiaFires_v2_2016") 
f17 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2017.shp", "SiberiaFires_v2_2017") 
f18 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2018.shp", "SiberiaFires_v2_2018") 
f19 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2019.shp", "SiberiaFires_v2_2019") 
f20 = st_read("../data/3.4-Vectors/SiberiaFires_v2_2020.shp", "SiberiaFires_v2_2020") 
```

```{r}
f20
```

```{r}
f20 %>% filter(IDobj == "3532")

```

# Combine all shapefiles
```{r}
all_fires = dplyr::bind_rows(f01, f02, f03, f04, f05, f06, f07, f08, f09, f10, f11, f12, f13, f14, f15, f16, f17, f18, f19, f20)
```

# Look at combined data
```{r}
head(all_fires)
```

```{r}
tail(all_fires)
```

```{r}
all_fires
```



# write to check data
```{r eval=FALSE, include=FALSE}
st_write(all_fires, "../outputs/3.5-AllFires/all_fires_2001-2020.shp", driver="ESRI Shapefile") 
```

# Final Cleaning

- Remove polygon that drew as large triangle
- Remove SSFS because it is not in the Boreal or Tundra Biome
- Remove fire smaller then 200 ha following Canadian large wildfire database
- Remove any fires with mean confidence of zero based on Modis hotspots
```{r}
all_fires1 = all_fires %>% 
  filter(!(FireYr == "2014" & IDobj=="2147")) %>%
  filter((EcoCode!="SSFS")) %>% 
  filter(SizeHa >= 200) %>%
  filter(AvgConf > 0) 
```



```{r}
all_fires1
```



# Check unique IDs
```{r}
all_fires1
```
22091
```{r}
n_distinct(all_fires1$FireId)
n_distinct(all_fires1$FID)
n_distinct(all_fires1$IDobj)
```

# Add Unique ID

```{r}
all_fires2 = all_fires1 %>% 
  mutate(UniqueId = random_id(n= 22091, bytes = 4)) %>%
  dplyr::select(UniqueId, IDobj:geometry)
```

```{r}
head(all_fires2)
```

```{r}
n_distinct(all_fires2$UniqueId)
```

# Write to shapefile
```{r eval=FALSE, include=FALSE}
st_write(all_fires2, "../outputs/3.5-AllFires/SiberiaFires2001-2020.shp", driver="ESRI Shapefile") 
```
